% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatorORFs.r
\name{annotatorORFs}
\alias{annotatorORFs}
\title{Annotate open reading frames}
\usage{
annotatorORFs(f.orf, f.fasta, f.gtf = NULL, f.prts, d.work = NULL)
}
\arguments{
\item{f.orf}{character string giving the name of tab-delimited TXT file with coordinates of open reading frames. This file should include three mandatory fields:
i) transcript_id - transcript ID;
ii) start        - start coordinate of open reading frame in a transcript;
iii) end         - end coordinate of open reading frame in a transcript.}

\item{f.fasta}{character string giving the name of FASTA file with nucleotide sequences of interest.}

\item{f.gtf}{character string giving the name of GTF/GFF file with transcripts of interest.
Default value is NULL.}

\item{f.prts}{character string giving the name of FASTA file with sequences of in sillico translated proteins.}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
data.frame object with columns:
 \item{transcript_id}{transcript ID.}
 \item{f_utr.length}{length of 5'UTRs.}
 \item{start.codon}{type of start codon.}
 \item{orf.start}{start coordinate of open reading frames.}
 \item{orf.stop}{stop coordinate of open reading frames.}
 \item{stop.codon}{type of stop codon.}
 \item{stop.status}{PTC status of stop codon.}
 \item{orf.length}{length of open reading frames.}
 \item{t_utr.length}{length of 3'UTRs.}
 \item{MW}{molecular weight.}
 \item{pI}{isoelectic point of a protein sequence.}
 \item{indexPPI}{potential protein interaction index.}
}
\description{
Annotate the identified open reading frames in nucleotide sequences of interest.
}
\examples{
f.orf_path <- system.file("extdata", "Set.trans_ORFs.coordinates.txt", 
                          package = "ORFhunteR")
f.fasta_path <- system.file("extdata", "Set.trans_sequences.fasta", 
                          package = "ORFhunteR")
f.gtf_path <- system.file("extdata", "Set.trans_sequences.gtf", 
                          package = "ORFhunteR")
f.prts_path <- system.file("extdata", "Set.trans_proteins.sequences.fasta", 
                          package = "ORFhunteR")
anno = annotatorORFs(f.orf = f.orf_path,
                     f.fasta = f.fasta_path,
                     f.gtf = f.gtf_path,
                     f.prts = f.prts_path,
                     d.work = NULL)
}
\author{
Vasily V. Grinev
}
