% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonStartStop.r
\name{codonStartStop}
\alias{codonStartStop}
\title{Identify all potential start and stop codons in a nucleotide sequence}
\usage{
codonStartStop(x)
}
\arguments{
\item{x}{character string giving the nucleotide sequence.}
}
\value{
list of potential start and stop codons with their coordinates.
}
\description{
This function scans a nucleotide sequence of interest
    in the search of start codon ATG or non-canonical start codons "CTG", 
    "GTG", "AGG", "TTG", "AAG", "ACG", "ATC", "ATT" or "ATA" (according to 
    Chothani S. P. et al. Mol Cell. 2022 Aug 4;82(15):2885-2899.e8) as well
    as stop codons TAA, TAG and TGA.
}
\examples{
codons <- codonStartStop(x="AAAATGGCATGGTAAGTCAAAATGGCATGGTAAGTCAAAATGGCGG")
}
\author{
Vasily V. Grinev
Last updated: June 12, 2024
}
