% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterEnsemblORFs.r
\name{filterEnsemblORFs}
\alias{filterEnsemblORFs}
\title{Filter of human Ensembl transcripts with inconsistent open reading frames}
\usage{
filterEnsemblORFs(f.fasta, f.mtRNAs, f.orf_coords, d.work = NULL)
}
\arguments{
\item{f.fasta}{character string giving the name of multi-FASTA file.}

\item{f.mtRNAs}{character string giving the name of TXT file in tab-delimited format containing
the list of mitochondrial transcripts. This file should include one mandatory
field transcript_id.}

\item{f.orf_coords}{character string giving the name of TXT file in tab-delimited format containing
coordinates of open reading frames for transcripts to be analyzed.
This file should include three mandatory fields:
i) transcript_id - transcript ID;
ii) start        - start coordinate of ORF;
iii) end         - end coordinate of ORF.}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
DNAStringSet object
}
\description{
Filter of human Ensembl transcripts with inconsistent open reading frames.
}
\examples{
f.fasta_path <- system.file("extdata", "Ensembl_release_85_GRCh38.p7_mRNAs_ex.fasta",
                           package = "ORFhunteR")
f.mtRNAs_path <- system.file("extdata", "Ensembl_release_85_GRCh38.p7_mitochondrial_RNAs.txt",
                           package = "ORFhunteR")
f.orf_coords_path <- system.file("extdata",
                           "Ensembl_release_85_GRCh38.p7_CDS_coordinates_of_mRNAs.txt",
                           package = "ORFhunteR")
res <- filterEnsemblORFs(f.fasta = f.fasta_path,
                         f.mtRNAs = f.mtRNAs_path,
                         f.orf_coords = f.orf_coords_path,
                         d.work = NULL)
# writeXStringSet(x = res, filepath = "filtered_mRNAs.fasta")
}
\author{
Vasily V. Grinev
}
