% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finderORFs.r
\name{finderORFs}
\alias{finderORFs}
\title{Identify all possible variants of open reading frames in a nucleotide sequence}
\usage{
finderORFs(x, cod.start = "ATG", f.orf = NULL, d.work = NULL)
}
\arguments{
\item{x}{character string giving the nucleotide sequence of interest.}

\item{cod.start}{character string with type of start codon: "ATG", "GTG", "TTG" or "CTG".
Default value is "ATG".}

\item{f.orf}{character string giving the name of output TXT file in tab-delimited format with
coordinates of identified open reading frames. Default value is NULL.}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
matrix with start and stop positions, length and sequence of identified variants of open reading frames.
}
\description{
Identify all possible variants of open reading frames in a nucleotide sequence of interest.
}
\examples{
x = "AAAATGGCTGCGTAATGCAAAATGGCTGCGAATGCAAAATGGCTGCGAATGCCGGCACGTTGCTACGT"
orf = finderORFs(x = x,
                cod.start = "ATG",
                f.orf = NULL,
                d.work = NULL)
}
\author{
Vasily V. Grinev
}
