% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finderPTCs.r
\name{finderPTC}
\alias{finderPTC}
\title{Identify the premature termination codons in nucleotide sequences}
\usage{
finderPTC(f.orf, f.trans, frt = "gtf", d.work = NULL)
}
\arguments{
\item{f.orf}{character string giving the name of TXT file in tab-delimited format with
coordinates of open reading frames.
This file should include four mandatory fields:
i) transcript_id - transcript ID;
ii) start        - start coordinate of open reading frame in a transcript;
iii) end         - end coordinate of open reading frame in a transcript;
iv) length       - length of open reading frame.}

\item{f.trans}{character string giving the name of file with transcript(-s) of interest.}

\item{frt}{format of file with transcript(-s) of interest. Default value is "gtf".}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
data.frame object with start and stop positions, length and stop status of codons for each transcript ID.
}
\description{
Identify the premature termination codons in nucleotide sequences of interest.
}
\examples{
f.orf_path <- system.file("extdata", "Set.trans_ORFs.coordinates.txt", package = "ORFhunteR")
f.trans_path <- system.file("extdata", "Set.trans_sequences.gtf", package = "ORFhunteR")
ptc <- finderPTC(f.orf = f.orf_path,
                 f.trans = f.trans_path,
                 frt = "gtf",
                 d.work = NULL)
}
\author{
Vasily V. Grinev
}
