% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaderTrExper.r
\name{loaderTrExper}
\alias{loaderTrExper}
\title{Load a set of transcripts}
\usage{
loaderTrExper(f.TrExper, genome = "BSgenome.Hsapiens.UCSC.hg38", d.work = NULL)
}
\arguments{
\item{f.TrExper}{character string giving the name of file with experimental transcripts. Allowed
file formats are "fasta", "fa", "gtf" or "gff".}

\item{genome}{character string giving the name of BSgenome data package
with full genome sequences. Default value is "BSgenome.Hsapiens.UCSC.hg38".}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
list of loaded transcript sequences.
}
\description{
Load a set of experimental transcripts.
}
\examples{
f.TrExper_path <- system.file("extdata", "Set.trans_sequences.gtf", package = "ORFhunteR")
tr.seq = loaderTrExper(f.TrExper = f.TrExper_path,
                       genome = "BSgenome.Hsapiens.UCSC.hg38",
                       d.work = NULL)
}
\author{
Vasily V. Grinev
}
