% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceORFs.R
\name{sequenceORFs}
\alias{sequenceORFs}
\title{Extract the sequences of identified open reading frames}
\usage{
sequenceORFs(f.orf, f.trans, frt = "fasta", f.seq = NULL, d.work = NULL)
}
\arguments{
\item{f.orf}{character string giving the name of tab-delimited TXT file with coordinates of open
reading frame(-s). This file should include three mandatory fields:
i) transcript_id - transcript ID;
ii) start        - start coordinate of open reading frame in a transcript;
iii) end         - end coordinate of open reading frame in a transcript.}

\item{f.trans}{character string giving the name of file with transcript(-s) of interest.}

\item{frt}{format of file with transcript(-s) of interest. Valid formats are "gtf", "gff",
"fasta" and "fa". Default value is "fasta".}

\item{f.seq}{character string giving the name of output FASTA file with sequences of identified
open reading frames. Default value is NULL.}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
DNAStringSet object with sequences of extracted open reading frames.
}
\description{
Extract the sequences of identified open reading frames.
}
\examples{
f.orf_path <- system.file("extdata", "Set.trans_ORFs.coordinates.txt", package = "ORFhunteR")
f.trans_path <- system.file("extdata", "Set.trans_sequences.fasta", package = "ORFhunteR")
seqORF = sequenceORFs(f.orf = f.orf_path,
                      f.trans = f.trans_path,
                      frt = "fasta",
                      f.seq = NULL,
                      d.work = NULL)
}
\author{
Vasily V. Grinev
}
