% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{translate}
\alias{translate}
\title{Translate any nucleotide sequences to peptide sequences}
\usage{
translate(
  x,
  codones = "internal",
  aaSymbol = 1,
  pep_length = 8,
  workDir = NULL
)
}
\arguments{
\item{x}{character string giving the path to FASTA file with nucleotide
sequences of interest.}

\item{codones}{location of tab-delimited TXT file with genetic code. Default
value is "internal" for system extdata. Alternatively, the name of
tab-delimited TXT file located in work directory should be specified.}

\item{aaSymbol}{type of amino acid symbols (one- or three-letter coding).
Default value is 1.}

\item{pep_length}{minimal length of in silico translated peptide/protein to
be saved. Default value is 8.}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that mean the current working directory.}
}
\value{
list of three AAStringSet objects with peptide sequences.
}
\description{
Translation of any nucleotide sequences to peptide sequences.
}
\examples{
seqs <- system.file("extdata",
                   "Set.trans_sequences.for.translation.fasta",
                   package="ORFhunteR")
res <- translate(x=seqs,
                 codones="internal",
                 aaSymbol=1,
                 pep_length=8,
                 workDir=NULL)
}
\author{
Vasily V. Grinev, Katsiaryna V. Huzava
Last updated: August 19, 2024.
}
