% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translatorORFs.r
\name{translatorORFs}
\alias{translatorORFs}
\title{Translate open reading frames to proteins}
\usage{
translatorORFs(f.seq_orf, aa.symbol = 1, f.prts = NULL, d.work = NULL)
}
\arguments{
\item{f.seq_orf}{character string giving the name of FASTA file with sequences of identified open
reading frames.}

\item{aa.symbol}{type of amino acid symbols (one- or three-letter coding). Default value is 1.}

\item{f.prts}{character string giving the name of output FASTA file with sequences of in sillico
translated proteins. Default value is NULL.}

\item{d.work}{character string giving the path to and name of work directory. Default value is NULL.}
}
\value{
AAStringSet object with sequences of obtained proteins.
}
\description{
Translate the identified open reading frames to proteins.
}
\examples{
f.seq_orf_path <- system.file("extdata", "Set.trans_ORFs.sequences.fasta", package = "ORFhunteR")
p.seq <- translatorORFs(f.seq_orf = f.seq_orf_path,
                        aa.symbol = 1)
}
\author{
Vasily V. Grinev
}
