% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorizeORFs.R
\name{vectorizeORFs}
\alias{vectorizeORFs}
\title{Vectorize the ORF(-s) sequence(-s) to sequence features}
\usage{
vectorizeORFs(x)
}
\arguments{
\item{x}{DNAStringSet object with ORF(-s) sequence(-s).}
}
\value{
The object of class data.frame
    with ORF(-s) vectorized into sequence features.
}
\description{
Vectorize the ORF(-s) sequence(-s) to sequence features.
}
\examples{
x = DNAStringSet(x = "ATGGGCCTCA")
feats <- vectorizeORFs(x = x)
}
\author{
Mikalai M. Yatskou
}
