% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllORFs.r
\name{findAllORFs}
\alias{findAllORFs}
\title{Find all variants of open reading frames in nucleotide sequence(-s)}
\usage{
findAllORFs(x, codStart = "ATG", workDir = NULL)
}
\arguments{
\item{x}{nucleotide sequence(-s) of interest which can be specified in the
following ways (see examples below):
i)   as character string of single nucleotide sequence;
ii)  as character vector of several nucleotide sequences;
iii) as R system file "Set.trans_sequences.fasta" with test set of
     nucleotide sequences;
iv)  as character string giving the name of file with custom sequences
     of interest. Usually it is a set of sequences of RNA molecules.
     This file must be located in working directory. Allowed file formats
     are "fasta" or "fa";
v)   as ready-to-use object of class DNAStringSet stored in the
     computer's RAM.}

\item{codStart}{character string with type of start codon (according to
Chothani S. P. et al. Mol Cell. 2022 Aug 4;82(15):2885-2899.e8): "ATG",
"CTG", "GTG", "AGG", "TTG", "AAG", "ACG", "ATC", "ATT" or "ATA".
Default value is "ATG".}

\item{workDir}{character string giving the path to and name of work
directory. NULL by default that means the current working directory.}
}
\value{
data frame with six following fields:
    i) transcript_id - ID of sequence;
    ii) orf_id       - ID of open reading frame;
    iii) start       - start coordinate of open reading frame in a sequence;
    iv) end          - end coordinate of open reading frame in a sequence;
    v) length        - length of open reading frame;
    vi) orf.sequence - sequence of open reading frame.
}
\description{
Identify all possible variants of open reading frames in
    nucleotide sequence(-s) of interest.
}
\examples{
### Input of nucleotide sequence of interest as character string:
tr <- "ACGCCTAGATGGTAGGGCCCATTTGCGAAGGACGAAGTGCCCAGTGAGTTAGCCAC"
orf <- findAllORFs(x=tr, codStart="ATG", workDir=NULL)
### Input of several nucleotide sequences of interest as character vector:
tr <- c("ACGCCTAGATGGTAGGGCCCATTTGCGAAGGACGAAGTGCCCAGTGAGTTAGCCAC",
        "GGGCCATGAGGGTAGCGCACATGGATGGACGATGTGCCCAGTGGGTTAGCCAC")
orf <- findAllORFs(x=tr, codStart="ATG", workDir=NULL)
### Input of R system file with test set of nucleotide sequences:
tr <- system.file("extdata",
                  "Set.trans_sequences.fasta",
                  package="ORFhunteR")
orf <- findAllORFs(x=tr, codStart="ATG", workDir=NULL)
### Input of custom file with sequences of interest:
# tr <- "path/Set.trans_sequences.fasta" 
# orf <- findAllORFs(x=tr, codStart="ATG", workDir=NULL)
### Usage of alternative start codon:
tr <- "ACGCCTAGCTGGATTTGCCCATATGCGATTTACGAAGTGCCCAGTGAGTTAGCCAC"
orf <- findAllORFs(x=tr, codStart="CTG", workDir=NULL)
}
\author{
Vasily V. Grinev
}
