% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicting_model.R
\name{predictORF}
\alias{predictORF}
\title{Predict the ORFs coordinates}
\usage{
predictORF(f.TrExper, genome = "BSgenome.Hsapiens.UCSC.hg38", d.work = NULL)
}
\arguments{
\item{f.TrExper}{character string giving the name of the data file.}

\item{genome}{character string giving the name of BSgenome data package
with full genome sequences. Default value is "BSgenome.Hsapiens.UCSC.hg38".}

\item{d.work}{character string giving the path to and name of work directory. Concatenated with filename via "/". Default value is NULL.}
}
\value{
The ORF coordinates of the RNA molecules.
}
\description{
Predict the ORFs coordinates and write it in txt file in your directory.
}
\examples{
\dontrun{
ORFs_coordinates <- predictORF(f.TrExper = "mRNAs_first200.fasta", d.work = NULL)
}
}
\author{
Mikalai M. Yatskou
}
